Fix placment of unsafe blocks
More often than not, entire functions are marked as unsafe, rather than
the exact place wher unsafe code is used. This not only makes it harder
to track possible UB, it also generates more code (see RFC 2585).

This commit changes all unsafe blocks in the `maa_sys` module to either
encompass the smallest logical unit. Alternatively, this commit also
adds unsafe checks to functions which only contain a single line of
unsafe code, or where this is not ppossible, marks the entire function
as unsafe.