Update .gitignore
The .gitignore file was missing some key files and directories. A proper
.giignore doesn't just exclude the `/target`` directory, but also
`/debug`. Furthermore, the `Cargo.lock` file is not supposed to be
committed to the repository. For a more in-depth reasoning see Chapter
2.6 in the Cargo Book:
https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html as
well as the FAQ chapter "Why have Cargo.lock in version control?" in the
same book:
https://doc.rust-lang.org/cargo/faq.html#why-have-cargolock-in-version-control

The new .gitignore will now also exclude backup files made by rustfmt
and MSVC debug files.