Change config file to use TOML instead of JSON
TOML files are usually the better choice for configuration files as they
are more human reasdable, shorter. TOML files support comments and, as
JSON, can be validated with a schema.
TOML files also avoid many of the pitfalls that the complex YAML
standard has, or some pitfalls of JSON where numbers are not IEEE 754
floating-point numbers.