Make version API return cargo.toml version
Hardcoding the version in a const field inside main.rs is error prone,
when the same version (should) also exist in the Cargo.toml file. We can
read that version in detail since the variables are set by cargo and
therefore we can get them with the `env!` and `option_env!` macros that
read the values during compile-time.