Add thiserror to workspace
Thiserror offers a simple and quick way to link multiple error types
together without just dumping them into an untyped `Box<dyn Error>`. It
is totally transparent after compilation. However, this is a temporary
measure and in the end this crate will be to be removed for custom
`From` and `Display` implementations (probably).