Make generated bindings a separate module
Up until now, the generated bindings were included as-is in the main FFI
module. This resulted in them being exported as part of the public API,
which is not ideal for a crate that povides a wrapper around the
aforementioned bindings. By making it its own separate module, it's
possible to control the visibility of all bindings.