Refactor the assistant construction code
In line with abstracting the ffi wrappers, the main construction code
for the assistant was changed. The callback now accepts _any_ Rust
function without leaking the generated C bindings. This allows the
backend to be changed, without impacting consumers.
The construction code has been condensed into a single `new` function,
which accepts an `Option<Callback>`. While this is a bit more verbose,
it also clearly shows that the callback is optional and if the user
wants to use it, even the default callback, they have to explicitly
state as such.

This commit also starts to introduce newtypes for various exported
types, mainly those that are used as IDs. While this intoduces a level
of indirection compared to plain typedef, it also makes sure that IDs,
which are (mostly) integer based, cannot be mixed up.