Change `load_resource` to accept a proper path
While in other languages it might be common to use strings as path
representations, Rust has multiple path types to work with and
functions dealing with paths should accept those. Howeverthis change
also introdues a potential performance regression on Windows, as there
is no way to simply convert a `Path` to a `CString` without taking the
in-between step of converting it to a `str` first. The reason for this
is Windows' internal path storage that allows invalid UTF-8 characters
to be used. While Microsoft has given the option to force the use of
UTF-8 in path APIs, this requres a minimum windows version that is
greater than the minimum versin MAA and Rust support. In the future,
this feature should be enabled.