Refactor connect method
The naming of the connect mehthod didn't use the the option to not block
on connect. This commit adds this option as a parameter and renames the
method to reflect that this method now is async.
Furthermore, the method previously accepted a string as target address.
However, the target address was always a socket address, which is
available as type in the standard library.
Lastly, the config parameter now accepts any type that implements
`Into<string>` and returns a newtype `AsyncCallId`.